hv: Initialize adb with client support only when necessary
Hierarchy Viewer uses the ddm protocol only when the ANDROID_HVPROTO
env var is set to "ddm". However, it was always initializing the
debug bridge with client support. This resulted in not being able
to run ddms when hv was already launched.

This CL updates hv to request client support only when ddm protocol
is used for communication.

Change-Id: Ie0d48e87d704c77aaae4dc9e1ce9767eb783887c