Fix Chimpchat ImageUtilsTest
The ImageUtilsTest tries to deserialize the image1.raw file from the
resourcesi but this fails since it tries to recreate the old
monkeyrunner RawImage class which has been replaced by the ChimRawImage
since then. The fix is just to reimplement the obsolete class as-is so that
deserialization can work properly.
Both RawImage implementations are compatible and interchangeable.

(cherry picked from commit a8764979fecc0b41e5525ded97bd4842a7556b48)

Change-Id: I90b7eacbb529f1450087228f7cec285d7cd14521