Fix Chimpchat ImageUtilsTest
The ImageUtilsTest tries to deserialize the image1.raw file from the
resourcesi but this fails since it tries to recreate the old
monkeyrunner RawImage class which has been replaced by the ChimRawImage
since then. The fix is just to reimplement the obsolete class as-is so that
deserialization can work properly.
Both RawImage implementations are compatible and interchangeable.

Change-Id: I927a21588944652c1823447706003fb432cc361a