Added an Evaluate Contrast button.
Evaluate Contrast tests the contrast between text color and background for
Views that may contain text. Sufficient contrast is defined as having a
contrast ratio of 4.5:1 in general, except for: Large text, which can have a
contrast ratio of only 3:1, Inactive components or pure decorations, and
text that is part of a logo or brand name.

Change-Id: Ib8bdb545e8407ff27531b743b2424784d5db52e5